Fix end vs complete flag use in deliver_events

+++ nestkernel/event_delivery_manager_impl.h
@@ -109,11 +109,8 @@ EventDeliveryManager::send_remote( thread tid, SpikeEvent& e, const long lag )
 
   for ( std::vector< Target >::const_iterator it = targets.begin(); it != targets.end(); ++it )
   {
-    // TODO: Check if we should re-activate assinging to threads
-    // const thread assigned_tid = ( *it ).get_rank() / kernel().vp_manager.get_num_assigned_ranks_per_thread();
-
     // Unroll spike multiplicity as plastic synapses only handle individual spikes.
-    for ( int i = 0; i < e.get_multiplicity(); ++i )
+    for ( size_t i = 0 ; i < e.get_multiplicity() ; ++i )
     {
       ( *emitted_spikes_register_[ tid ] )[ lag ].push_back( *it );
     }
